some optimizations

+++ src/graphics_new/zbuff/reset_zbuff.c
@@ -6,7 +6,7 @@
 /*   By: mfischer <mfischer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/09 17:55:53 by mfischer          #+#    #+#             */
-/*   Updated: 2019/06/10 20:00:59 by mfischer         ###   ########.fr       */
+/*   Updated: 2019/06/21 19:33:18 by mfischer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 void	reset_zbuff(double value, size_t size)
 {
-	int		i;
+	size_t		i;
 	double	*buff;
 
 	buff = get_zbuff();
 	i = -1;
-	while ((size_t)++i < size)
+	while (++i < size)
 		buff[i] = value;
 }
